product_manager/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
pydantic==2.7.4
python-dotenv==1.0.1
reportlab==4.2.2
alembic==1.13.1
httpx==0.27.0
